refactor: resettle of test_rederer
